Fix deopt & remove unneeded es5 shims.
It would be enough to build an array of all arguments, but we can
optimize further knowing that no official Intl constructors take
more than two arguments, and IntlMessageFormat takes 3.

This both prevents an array creation but also allows us to avoid
the bind shim to create the format object.

As this function is called very often it is helpful that is fast
and especially that it can be optimized.

We can also avoid the Object.create() shim as we know that
no cache key can possibly collide with a built-in, as
JSON.stringify(cacheId) will always return a key like "[...".